Refactor Expo native view hosting
