+++ packages/expo/ios/ClerkUserButtonNativeView.swift
@@ -1,35 +1,13 @@
 import UIKit
 
-public class ClerkUserButtonNativeView: UIView {
-  private lazy var hostingCoordinator = ClerkNativeHostingCoordinator(containerView: self)
-  private var hasInitialized: Bool = false
+public class ClerkUserButtonNativeView: ClerkNativeViewHost {
+  override func makeHostedController() -> UIViewController? {
+    guard let factory = clerkViewFactory else { return nil }
 
-  override public init(frame: CGRect) {
-    super.init(frame: frame)
+    return factory.createUserButton()
   }
 
-  public required init?(coder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-
-  override public func didMoveToWindow() {
-    super.didMoveToWindow()
-    if window != nil && !hasInitialized {
-      hasInitialized = true
-      updateView()
-    }
-  }
-
-  private func updateView() {
-    guard let factory = clerkViewFactory else { return }
-
-    guard let returnedController = factory.createUserButton() else { return }
-
-    hostingCoordinator.attach(returnedController, clearBackground: true)
-  }
-
-  override public func layoutSubviews() {
-    super.layoutSubviews()
-    hostingCoordinator.layout()
+  override var clearsHostedViewBackground: Bool {
+    true
   }
 }
